Edited Booking Table in DB, and made edits to DAO and some hibernate xmls
Changed Booking_timestamp in db to timestamp, and start date and end date to DateTime, as time can't seem to be displayed in the DB.

Also, hibernate is confusing. Although I set dateTime in db, it generates the xml as timestamp for start and end date. And, the POJOs generated are of Date type. :S confused X1000

Anyway, dao is up and running, i'm using testdb class to test CRUD.